# catalog/lease: adopt RWLock for descriptor states