Previously, the lease manager used regular mutexes for accessing
descriptor information. When running a simple workload like
BenchmarkLeaseManager we can see high contention on the descriptor state
locks during both acquire and release. Since, the operations are mainly
reads on this object, we can switch to a RWLock, since modification
only occurs with schema changes. This patch updates findForTimeStamp and
release to use read only locks for the hot code paths.

This produces the following improvement in the leasing micro-benchmark:
```
name                                           old time/op    new time/op    delta
LeaseManager/lease_benchmark_(threads=1)-10      21.1µs ± 1%    21.1µs ± 1%     ~     (p=0.968 n=9+10)
LeaseManager/lease_benchmark_(threads=128)-10    7.63ms ± 4%    8.13ms ± 6%   +6.58%  (p=0.003 n=10+10)
LeaseManager/lease_benchmark_(threads=64)-10     3.66ms ± 5%    4.00ms ± 6%   +9.30%  (p=0.000 n=10+10)
LeaseManager/lease_benchmark_(threads=32)-10     1.67ms ± 6%    1.93ms ± 6%  +15.76%  (p=0.000 n=10+10)
LeaseManager/lease_benchmark_(threads=16)-10      703µs ± 9%     886µs ± 9%  +25.98%  (p=0.000 n=10+10)

name                                           old alloc/op   new alloc/op   delta
LeaseManager/lease_benchmark_(threads=1)-10        144B ±80%      118B ±17%     ~     (p=0.128 n=9+10)
LeaseManager/lease_benchmark_(threads=16)-10    5.21kB ±114%    7.99kB ±75%     ~     (p=0.055 n=10+10)
LeaseManager/lease_benchmark_(threads=32)-10     6.15kB ±29%    5.62kB ±39%     ~     (p=0.280 n=10+10)
LeaseManager/lease_benchmark_(threads=128)-10    29.5kB ±19%    27.2kB ± 9%     ~     (p=0.234 n=8+8)
LeaseManager/lease_benchmark_(threads=64)-10      406kB ± 7%     447kB ±14%  +10.14%  (p=0.002 n=9+10)

name                                           old allocs/op  new allocs/op  delta
LeaseManager/lease_benchmark_(threads=1)-10       0.40 ±400%      0.00          ~     (p=0.211 n=10+10)
LeaseManager/lease_benchmark_(threads=16)-10      27.2 ±121%      41.8 ±79%     ~     (p=0.071 n=10+10)
LeaseManager/lease_benchmark_(threads=32)-10       17.1 ±35%      17.9 ± 6%     ~     (p=0.092 n=10+9)
LeaseManager/lease_benchmark_(threads=128)-10       198 ±18%       200 ± 4%     ~     (p=0.483 n=8+8)
LeaseManager/lease_benchmark_(threads=64)-10      1.58k ± 4%     1.74k ±10%   +9.57%  (p=0.000 n=10+10)
```

Fixes: #139166

Release note: None